change INT to Double

+++ app/src/main/java/com/product/nutriwise/data/local/preference/profile/ProfileModel.kt
@@ -3,7 +3,7 @@ package com.product.nutriwise.data.local.preference.profile
 data class ProfileModel(
     val umur : Int,
     val gender: Boolean,
-    val tinggibadan: Int,
-    val beratbadan: Int,
+    val tinggibadan: Double,
+    val beratbadan: Double,
     val aktivitas: Int
 )
